rewrote EmaliValidation in Swift
