--- Codemine/NSString+EmailValidation.h
@@ -1,15 +1,0 @@
-//
-//  NSString+EmailValidation.h
-//  
-//
-//  Created by Kasper Welner on 12/5/12.
-//  Copyright (c) 2012 Nodes. All rights reserved.
-//
-
-#import <Foundation/Foundation.h>
-
-@interface NSString (EmailValidation)
-
-- (BOOL)isValidEmailAddress;
-
-@end
