+++ Codemine/SharedExtensions.swift
@@ -28,25 +28,4 @@ public protocol NibInstantiable: class {
     static func newFromNib<T>() -> T
 }
 
-public extension String {
-    public func camelCaseToUnderscore() -> String {
-        var returnString = self
 
-        let characterArray = Array(returnString.characters).map { (character) -> String in
-            let inputCharacterString = String(character)
-            let lowerCaseCharacterString = String(character).lowercaseString
-
-            if inputCharacterString != lowerCaseCharacterString {
-                return "_" + lowerCaseCharacterString
-            }
-
-            return inputCharacterString
-        }
-
-        returnString = characterArray.reduce("") {
-            return $0 + $1
-        }
-
-        return returnString
-    }
-}
